backend/requirements.txt
requests
bs4
python-dotenv
fastapi
uvicorn
pydantic